Create BLEKeyboardPS2.ino
this keyboard library is meant for using the PS/2 port thats connected to a esp32 using FabGL libraries and scancode. keyboard arrows letters and numbers all work, crtl and F1 copy to and F2 paste the clipboard F3 F4 vol up down works. this code was tested on TTGO VGA V1.4 Controller PS/2 port. there are a some issues that need to be worked out on the timing. the F keys can all be programed to have macro functions. main issues is letter doubling, speed tweeks. control acts as the VK switch key and have to use it to use enter for some reason. enter can be programed into a F key as a workaround. not all characters features work on number pad. this code could be redone without FabGL to use PS/2 port and be much more precise with button up and down pushes.